% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmmc_main.R
\name{get_chain.cpmmc}
\alias{get_chain.cpmmc}
\title{S3 Implementatio of get_chain generic method for cpmmc}
\usage{
\method{get_chain}{cpmmc}(object, chain)
}
\arguments{
\item{chain}{type of chain wanted from "proposed", "accepted", "theta", "u", "proposed theta", "proposed u", "accepted theta", "accepted u"}

\item{cpmmc}{object}
}
\value{
chain="proposed" returns the proposed chain of theta and u

chain="accepted" returns the accepted chain of theta and u

chain="proposed theta" returns the proposed chain of theta

chain="proposed u" returns the proposed chain of u

chain="accepted theta" returns the accepted chain of theta

chain="accepted u" returns the accepted chain of u
}
\description{
Accesses the proposed and/or accepted chains for cpmmc object
}
