% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain_generic_functions.R
\name{get_chain}
\alias{get_chain}
\title{Generic function to access proposed and/or accepted chains from output of run_mh() function}
\usage{
get_chain(object, chain)
}
\arguments{
\item{object, }{markov chain object e.g. cpmmc}
}
\value{
list of parameters of latest accepted state
}
\description{
Generic function to access proposed and/or accepted chains from output of run_mh() function
}
