% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{get_chain_length.markov_chain}
\alias{get_chain_length.markov_chain}
\title{S3 Implementation of get_chain_length generic method for markov_chain}
\usage{
\method{get_chain_length}{markov_chain}(object)
}
\arguments{
\item{object, }{markov_chain object e.g. cpmmc, metropolis_hastings}
}
\value{
numeric value for the length of the Markov chain
}
\description{
S3 Implementation of get_chain_length generic method for markov_chain
}
