% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmmc_main.R
\name{get_state.cpmmc}
\alias{get_state.cpmmc}
\title{S3 Implementatio of get_state generic method for cpmmc}
\usage{
\method{get_state}{cpmmc}(object)
}
\arguments{
\item{cpmmc}{object}
}
\value{
list, parameters of latest state
}
\description{
Accesses the latest state of markov chain for cpmmc object
}
