% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{get_state.markov_chain}
\alias{get_state.markov_chain}
\title{S3 Implementation of get_state generic method for markov_chain}
\usage{
\method{get_state}{markov_chain}(object)
}
\arguments{
\item{object, }{markov_chain object e.g. cpmmc, metropolis_hastings}
}
\value{
list of parameters of the latest accepted state
}
\description{
S3 Implementation of get_state generic method for markov_chain
}
