% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{markov_chain}
\alias{markov_chain}
\title{Instantiate Markov chain}
\usage{
markov_chain(initial_value, transition_kernel)
}
\arguments{
\item{intial_value, }{numeric for initial state of chain, list for multiple named params}

\item{transition_sampler, }{function, takes current state and returns next state in chain}
}
\value{
markov_chain object
}
\description{
This is an S3 object constructor helper function for class random_effects_model which inherics from cpmmc class
}
