% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_hastings.R
\name{metropolis_hastings}
\alias{metropolis_hastings}
\title{Instantiate metropolis_hastings object and set initial parameters}
\usage{
metropolis_hastings(initial_params, log_target_density,
  log_proposal_density, proposal_sampler)
}
\arguments{
\item{initial_params, }{numeric for initial state of chain, list for multiple named parameters}

\item{log_target_density, }{the log target density for the MH algorithm, i.e. the log density for the stationary distribution of the Markov Chain}

\item{log_proposal_density, }{log density function for parameter proposals}

\item{proposal_sampler, }{distribution function to sample new parameter proposals}
}
\value{
metropolis_hastings object
}
\description{
This is an S3 object constructor helper function for class metropolis_hastings which inherit from markov_chain class
}
