% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_dist_helpers.R
\name{normal_density}
\alias{normal_density}
\title{Helper function to generate samplesfor univariate and multivariate normal distribution}
\usage{
normal_density(data, mean = NA, sigma = NA, ...)
}
\arguments{
\item{mean, }{mean of normal distribution}

\item{sigma, }{standard deviation (for univariate distribution), covariance matrix (for multivariate)}

\item{n, }{number of data points to sample
at each data point is the possibly multi-dimensional data}
}
\value{
numeric value, density
}
\description{
Samples normal distribution
}
