% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects_model.R
\name{normal_log_IS_estimate}
\alias{normal_log_IS_estimate}
\title{Helper function to perform importance sampling for the random effect model}
\usage{
normal_log_IS_estimate(data, theta, new_u)
}
\arguments{
\item{data, }{a array of data points, indexed first by the number of data points,
at each data point is the possibly multi-dimensional data}

\item{theta, }{a possibly multivariate array constant added to each 2-dimensional entry of u_t
must be same number of parameters as each entry of u_t and of each data-point}

\item{new_u, }{a 3-dimensional array of auxilliary variables used for proposal in importance sampling
dim = c(number_of_samples, dimension_of_each_point, number_of_data_points)}
}
\value{
numeric value, log of importance sampling estimate for g_theta = N(x,1), f_theta = N(theta,1)
}
\description{
Runs importance sampling
}
