% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_dist_helpers.R
\name{normal_sampler}
\alias{normal_sampler}
\title{Helper function to perform idensity calculation for univariate and multivariate normal distribution}
\usage{
normal_sampler(n, mean = NA, sigma = NA, ...)
}
\arguments{
\item{mean, }{mean of normal distribution}

\item{sigma, }{standard deviation (for univariate distribution), covariance matrix (for multivariate)}

\item{data, }{a array of data points, indexed first by the number of data points,
at each data point is the possibly multi-dimensional data}
}
\value{
numeric value, density
}
\description{
Calculates density
}
