% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{run_chain.markov_chain}
\alias{run_chain.markov_chain}
\title{S3 Implementation of run_chain generic method for markov_chain}
\usage{
\method{run_chain}{markov_chain}(object, chain_length)
}
\arguments{
\item{object, }{markov_chain object}

\item{chain_length, }{number of new states generated using the transition kernel}
}
\value{
markov_chain object, with chains of parameters updated according to the length of chain
}
\description{
Generate a Markov chain of required length using the transition kernel of the markov_chain object
}
