% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{run_chain}
\alias{run_chain}
\title{Generic function to extend the markov chain by a single step}
\usage{
run_chain(object, chain_length)
}
\arguments{
\item{object, }{markov chain object e.g. cpmmc}
}
\value{
object, markkov chain object with MC extended by n steps
}
\description{
Generic function to extend the markov chain by a single step
}
