% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain_generic_functions.R
\name{run_mh}
\alias{run_mh}
\title{Generic function to run Metropolis Hastings algorithm}
\usage{
run_mh(object, nsim)
}
\arguments{
\item{object, }{markov chain object e.g. cpmmc}

\item{N, }{number of samples from posterior}
}
\value{
list, proposal state and accepted state
}
\description{
Generic function to run Metropolis Hastings algorithm
}
