% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain_generic_functions.R
\name{single_mh_step}
\alias{single_mh_step}
\title{Generic function to run single Metropolis Hastings Step}
\usage{
single_mh_step(object)
}
\arguments{
\item{object, }{markov chain object e.g. cpmmc}
}
\value{
list, proposal state and accepted state
}
\description{
Generic function to run single Metropolis Hastings Step
}
