% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{single_transition.markov_chain}
\alias{single_transition.markov_chain}
\title{S3 Implementation of single_transition generic method for markov_chain}
\usage{
\method{single_transition}{markov_chain}(object)
}
\arguments{
\item{object, }{markov_chain object}
}
\value{
markov_chain object, with chains of parameters updated by one step using the transition kernel
}
\description{
Generate the next state of the Markov chain using the transition kernel
}
