% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{single_transition}
\alias{single_transition}
\title{Generic function to extend the markov chain by a single step}
\usage{
single_transition(object)
}
\arguments{
\item{object, }{markov_chain object e.g. cpmmc, metropolis_hastings}
}
\value{
object, markov_chain object with MC extended by 1 steps using MCMC algorithms specified by the class of the object
}
\description{
Generic function to extend the markov chain by a single step
}
