% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmmc_main.R
\name{cpmmc}
\alias{cpmmc}
\title{Instantiate cpmmc object and set initial parameters}
\usage{
cpmmc(data, theta_0, u_0, rho, marginal_estimator_func, theta_prior_density,
  theta_proposal_density, theta_proposal_sampler)
}
\value{
cpmmc object
}
\description{
This is an S3 object constructor helper function for class cpmmc
}
