% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{get_chain_length}
\alias{get_chain_length}
\title{Generic function to access proposed and/or accepted chains from output of run_mh() function}
\usage{
get_chain_length(object)
}
\arguments{
\item{object, }{markov chain object e.g. cpmmc}
}
\value{
list of parameters of latest accepted state
}
\description{
Generic function to access proposed and/or accepted chains from output of run_mh() function
}
