% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain_generic_functions.R
\name{get_state}
\alias{get_state}
\title{Generic function to access latest state in a markov chain}
\usage{
get_state(object)
}
\arguments{
\item{object, }{markov chain object e.g. cpmmc}
}
\value{
list of parameters of latest accepted state
}
\description{
Generic function to access latest state in a markov chain
}
