% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects_model.R
\name{normal_random_effect_model}
\alias{normal_random_effect_model}
\title{Instantiate random_effects_model object}
\usage{
normal_random_effect_model(data, theta_0, u_0, rho,
  log_theta_prior_density = function(x) dnorm(x, log = T),
  log_theta_proposal_density = function(old_theta, new_theta)
  dnorm(new_theta - old_theta, log = T),
  theta_proposal_sampler = function(theta) rnorm(1, mean = theta))
}
\value{
random_effects_model object
}
\description{
This is an S3 object constructor helper function for class random_effects_model which inherics from cpmmc class
}
