% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmmc_main.R
\name{run_mh.cpmmc}
\alias{run_mh.cpmmc}
\title{S3 Implementatio of run_mh generic method for cpmmc}
\usage{
\method{run_mh}{cpmmc}(object, N)
}
\arguments{
\item{N}{number of samples from posterior}

\item{cpmmc}{object}
}
\value{
object, with chains of parameters updated
}
\description{
Runs Metropolis Hastings Algorithm to generate N new proposals and accepted states from
 the latest state of markov chain for cpmmc object, using intrinsic cpmmc proposals
}
