% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmmc_main.R
\name{single_transition.cpmmc}
\alias{single_transition.cpmmc}
\title{S3 Implementatio of single_mh_step generic method for cpmmc}
\usage{
\method{single_transition}{cpmmc}(object)
}
\arguments{
\item{cpmmc}{object}
}
\value{
object, with chains of parameters updated
}
\description{
Runs Metropolis Hastings Algorithm to generate a new proposal and accepted state from
 the latest state of markov chain for cpmmc object, using intrinsic cpmmc proposals
}
