% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chain.R
\name{single_transition}
\alias{single_transition}
\title{Generic function to extend the markov chain by a single step}
\usage{
single_transition(object)
}
\arguments{
\item{object, }{markov chain object e.g. cpmmc}
}
\value{
object, markkov chain object with MC extended
}
\description{
Generic function to extend the markov chain by a single step
}
